added changes related to disk mounting

+++ hdfs2gcs-terraform/nifi-ca.tf
@@ -38,37 +38,39 @@ resource "google_compute_instance" "nifi-ca" {
     }
 
     metadata_startup_script =   <<EOF
+        if [[ ! -f /opt/startup-script-finished.txt ]]
+        then 
+            if [[ "${var.image}" == *"centos"* ]]; then
+            gsutil cp  ${var.nifi-bucket}/binaries/unzip*.rpm /opt
+            yum install  /opt/unzip*.rpm -y
+            else 
+            gsutil cp  ${var.nifi-bucket}/binaries/unzip*.deb /opt
+            apt-get install /opt/unzip*.deb -y
+            fi
+                
+            NIFI_UID=10000
+            NIFI_GID=10000
 
-        if [[ "${var.image}" == *"centos"* ]]; then
-          gsutil cp  ${var.nifi-bucket}/binaries/unzip*.rpm /opt
-          yum install  /opt/unzip*.rpm -y
-        else 
-          gsutil cp  ${var.nifi-bucket}/binaries/unzip*.deb /opt
-          apt-get install /opt/unzip*.deb -y
+            groupadd -g $${NIFI_GID} nifi || groupmod -n nifi `getent group $${NIFI_GID} | cut -d: -f1` \
+                && useradd --shell /bin/bash -u $${NIFI_UID} -g $${NIFI_GID} -m nifi \
+                && mkdir -p ${var.nifi-path} \
+                
+            mkdir -p /usr/lib/jvm/tmp-jdk
+            gsutil -m cp -r  ${var.nifi-bucket}/binaries/${var.jdkpackage} /usr/lib/jvm/
+            cd /usr/lib/jvm/ && tar -xzvf ${var.jdkpackage} -C /usr/lib/jvm/
+            rm -f /usr/lib/jvm/${var.jdkpackage}
+            cp -R /usr/lib/jvm/jdk*/* /usr/lib/jvm/tmp-jdk && rm -R -f /usr/lib/jvm/jdk* && mv /usr/lib/jvm/tmp-jdk /usr/lib/jvm/jdk 
+            chmod -R a+x  /usr/lib/jvm/
+            chown -R nifi:nifi /usr/lib/jvm/
+            echo "export JAVA_HOME=/usr/lib/jvm/jdk" >> ~/.bashrc
+            echo "export PATH=$PATH:/usr/lib/jvm/jdk/bin" >> ~/.bashrc
+            gsutil cp ${var.nifi-bucket}/binaries/nifi-toolkit-${var.nifi-version}-bin.zip ${var.nifi-path}
+            unzip ${var.nifi-path}/nifi-toolkit-${var.nifi-version}-bin.zip -d ${var.nifi-path}
+            rm ${var.nifi-path}/nifi-toolkit-${var.nifi-version}-bin.zip
+            chown nifi:nifi -R ${var.nifi-path}/*
+            find ${var.nifi-path} -type f -iname "*.sh" -exec chmod +x {} \;
+            touch /opt/startup-script-finished.txt && echo "the startup script run once" > /opt/startup-script-finished.txt
         fi
-              
-        NIFI_UID=10000
-        NIFI_GID=10000
-
-        groupadd -g $${NIFI_GID} nifi || groupmod -n nifi `getent group $${NIFI_GID} | cut -d: -f1` \
-            && useradd --shell /bin/bash -u $${NIFI_UID} -g $${NIFI_GID} -m nifi \
-            && mkdir -p ${var.nifi-path} \
-            
-        mkdir -p /usr/lib/jvm/tmp-jdk
-        gsutil -m cp -r  ${var.nifi-bucket}/binaries/${var.jdkpackage} /usr/lib/jvm/
-        cd /usr/lib/jvm/ && tar -xzvf ${var.jdkpackage} -C /usr/lib/jvm/
-        rm -f /usr/lib/jvm/${var.jdkpackage}
-        cp -R /usr/lib/jvm/jdk*/* /usr/lib/jvm/tmp-jdk && rm -R -f /usr/lib/jvm/jdk* && mv /usr/lib/jvm/tmp-jdk /usr/lib/jvm/jdk 
-        chmod -R a+x  /usr/lib/jvm/
-        chown -R nifi:nifi /usr/lib/jvm/
-        echo "export JAVA_HOME=/usr/lib/jvm/jdk" >> ~/.bashrc
-        echo "export PATH=$PATH:/usr/lib/jvm/jdk/bin" >> ~/.bashrc
-        gsutil cp ${var.nifi-bucket}/binaries/nifi-toolkit-${var.nifi-version}-bin.zip ${var.nifi-path}
-        unzip ${var.nifi-path}/nifi-toolkit-${var.nifi-version}-bin.zip -d ${var.nifi-path}
-        rm ${var.nifi-path}/nifi-toolkit-${var.nifi-version}-bin.zip
-        chown nifi:nifi -R ${var.nifi-path}/*
-        find ${var.nifi-path} -type f -iname "*.sh" -exec chmod +x {} \;
-
         su nifi -c 'export PATH=$PATH:/usr/lib/jvm/jdk/bin && cd /home/nifi && ${var.nifi-path}/nifi-toolkit-${var.nifi-version}/bin/tls-toolkit.sh server -c ${var.nifi-ca-hostname} -t ${var.ca-token} &'
         sleep 10
         cd /root
